more games

+++ public/assets/js/games.js
@@ -62,6 +62,10 @@ document.addEventListener('DOMContentLoaded', function () {
                 gameCard.appendChild(titleContainer);
 
                 gameCard.addEventListener('click', () => {
+                    if (game.url.includes('play.geforcenow.com')) {
+                        window.open(game.url, '_blank');
+                        return;
+                    }
                     if (game.url.includes('https://') || game.url.includes('http://') || game.url.includes('/assets/games/')) {
                         localStorage.setItem('url', game.url);
                     }
